superapp/apps/authentication/requirements.txt
django-guardian==2.4.*
